Extract code into files
