+++ packages/react-dev-utils/package.json
@@ -19,6 +19,7 @@
     "eslintFormatter.js",
     "errorOverlayMiddleware.js",
     "FileSizeReporter.js",
+    "forceUpdateHook.js",
     "printBuildError.js",
     "formatWebpackMessages.js",
     "getProcessForPort.js",
@@ -28,6 +29,7 @@
     "launchEditor.js",
     "launchEditorEndpoint.js",
     "ModuleScopePlugin.js",
+    "nodeHighlighter.js",
     "noopServiceWorkerMiddleware.js",
     "openBrowser.js",
     "openChrome.applescript",
